perf: implement filter
- Add DAR only filter(walkdir)
- Summarize common codes
