+++ cspell.jsonc
@@ -11,19 +11,23 @@
     "jwalk",
     "Mincho",
     "mohidden",
+    "mpsc",
     "multispace",
     "noconflict",
+    "println",
     "rfind",
     "Rustup",
     "Shippori",
     "Skyrim",
     "splitn",
     "stperson",
+    "struct",
     "tauri",
     "thiserror",
     "Unhide",
     "unlisten",
     "unoptimized",
+    "Usize",
     "walkdir",
     "Warhammer",
     "worldspace"
