+++ dar2oar_core/Cargo.toml
@@ -11,7 +11,6 @@ rust-version = "1.60"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow = { version = "1.0.75", features = ["backtrace"] }
 async-walkdir = "0.2.0" # To traverse dir recursivly(Single thiread)
 jwalk = "0.8.1" # To parallel traverse dir recursivly
 nom = { version = "7.1.3", features = ["alloc"] } # Syntax
@@ -28,14 +27,16 @@ tokio-stream = "0.1.14" # Async next() method
 tracing = "0.1.40" # Logger
 
 [dev-dependencies]
+anyhow = { version = "1.0.75", features = ["backtrace"] }
 criterion = { version = "0.5.1", features = [
   "async_futures",
   "async_tokio",
   "html_reports",
 ] }
 once_cell = "1.18.0"
-tracing-appender = "0.2"
 pretty_assertions = "1.4.0"
+temp-dir = "0.1.11"
+tracing-appender = "0.2"
 tracing-subscriber = "0.3.17"
