--- dar2oar_core/src/fs/async_closure.rs
@@ -1,4 +1,0 @@
-pub struct AsyncClosure;
-impl AsyncClosure {
-    pub async fn default(_: usize) {}
-}
