+++ dar2oar_core/src/fs/mapping_table.rs
@@ -1,3 +1,4 @@
+use crate::error::Result;
 use std::collections::HashMap;
 use std::path::Path;
 use tokio::{fs::File, io::AsyncReadExt};
@@ -13,9 +14,7 @@ pub async fn get_mapping_table(
 }
 
 /// Try to read mapping table from path
-pub async fn read_mapping_table(
-    table_path: impl AsRef<Path>,
-) -> anyhow::Result<HashMap<String, String>> {
+pub async fn read_mapping_table(table_path: impl AsRef<Path>) -> Result<HashMap<String, String>> {
     let mut file_contents = String::new();
     File::open(table_path)
         .await?
