+++ src-tauri/src/cmd.rs
@@ -1,8 +1,5 @@
 use crate::convert_option::{AsyncFrom, GuiConverterOptions};
-use dar2oar_core::{
-    convert_dar_to_oar,
-    fs::{async_closure::AsyncClosure, parallel, remove_oar, unhide_dar, ConvertOptions},
-};
+use dar2oar_core::{convert_dar_to_oar, remove_oar, unhide_dar, Closure, ConvertOptions};
 use tauri::Window;
 
 /// logger hook or bail!
@@ -29,14 +26,8 @@ macro_rules! bail {
 #[tauri::command]
 pub(crate) async fn convert_dar2oar(options: GuiConverterOptions<'_>) -> Result<String, String> {
     tracing::debug!("options: {:?}", &options);
-    let run_parallel = options.run_parallel.unwrap_or_default();
     let config = ConvertOptions::async_from(options).await;
-
-    let res = match run_parallel {
-        true => parallel::convert_dar_to_oar(config, AsyncClosure::default).await,
-        false => convert_dar_to_oar(config, AsyncClosure::default).await,
-    };
-    response!(res)
+    response!(convert_dar_to_oar(config, Closure::default).await)
 }
 
 #[tauri::command]
@@ -45,7 +36,6 @@ pub(crate) async fn convert_dar2oar_with_progress(
     options: GuiConverterOptions<'_>,
 ) -> Result<String, String> {
     tracing::debug!("options: {:?}", &options);
-    let run_parallel = options.run_parallel.unwrap_or_default();
 
     #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
     struct Payload {
@@ -55,29 +45,16 @@ pub(crate) async fn convert_dar2oar_with_progress(
         if let Err(err) = window.emit("show-progress", Payload { index }) {
             tracing::error!("{}", err);
         };
-        async move {}
     };
     let config = ConvertOptions::async_from(options).await;
 
-    let res = match run_parallel {
-        true => parallel::convert_dar_to_oar(config, sender).await,
-        false => convert_dar_to_oar(config, sender).await,
-    };
-    response!(res)
+    response!(convert_dar_to_oar(config, sender).await)
 }
 
 #[tauri::command]
 pub(crate) async fn change_log_level(log_level: Option<&str>) -> Result<(), String> {
     tracing::debug!("Selected log level: {:?}", log_level);
-    let log_level = log_level.unwrap_or("error");
-    let log_level = match log_level {
-        "trace" | "debug" | "info" | "warn" | "error" => log_level,
-        unknown_level => {
-            tracing::warn!("Unknown log level {}. Fallback to error", unknown_level);
-            "error"
-        }
-    };
-    crate::logging::change_log_level(log_level).or_else(|err| bail!(err))
+    crate::logging::change_log_level(log_level.unwrap_or("error")).or_else(|err| bail!(err))
 }
 
 #[tauri::command]
